fix(wave1): foundation cleanup — StrEnum dedup, HTMX path normalize, dead code removal
- Extract StrEnum backport to engine/core/_compat.py (was duplicated in types.py, events.py, permissions.py)
- Normalize HTMX producer paths to /api/v1/ in producers.html (was mixing /api/ and /api/v1/)
- Delete dead dashboard/templates/config.html (page redirects to /settings)
- Remove unreachable level<0/level>4 check in kill_switch.py (Pydantic Field already validates)
